% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalprojectHK.R
\name{error}
\alias{error}
\title{type I error rate control}
\usage{
error(alpha, lambda0, B, n)
}
\arguments{
\item{alpha}{Nominal type I error rate α}

\item{lambda0}{λ0 value in Poisson distribution}

\item{B}{Number of repeats}

\item{n}{Limited sample size}
}
\description{
\code{error} output type I error rate control of Wald, Score, and Likelihood ratio test for Poisson distribution with specified sample size with bootstrapping method
}
\examples{
# error(0.05, 5, 100, 20)
# error(0.05, 10, 1000, 2000)
}
