% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalprojectHK.R
\name{huber}
\alias{huber}
\title{Huber function}
\usage{
huber(x)
}
\arguments{
\item{x}{numeric value of x axis}
}
\description{
\code{huber} output Huber function value from x axis value
}
\examples{
# huber(1)
# huber(c(1,2,3))
}
