% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalprojectHK.R
\name{power}
\alias{power}
\title{Statistical power of Wald, Score, and Likelihood ratio test in specified sample size}
\usage{
power(alpha, lambda0, lambda1, B, n)
}
\arguments{
\item{alpha}{Nominal type I error rate α}

\item{lambda0}{λ0 value in Poisson distribution}

\item{lambda1}{λ1 value in Poisson distribution}

\item{B}{Number of repeats}

\item{n}{Limited sample size}
}
\description{
\code{power} Output statistical power of Wald, Score, and Likelihood ratio test for Poisson distribution with specified sample size with bootstrapping method
}
\examples{
# power(0.05, 4, 5, 100, 20)
# power(0.05, 9, 10, 1000, 2000)
}
