% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalprojectHK.R
\name{rounding}
\alias{rounding}
\title{Rounding to a specified decimal point}
\usage{
rounding(x, d)
}
\arguments{
\item{x}{numeric value which want to be rounded}

\item{d}{decimal points to be rounded}
}
\description{
\code{rounding} Rounding to a specified decimal point
}
\examples{
# rounding(2.55, 1)

}
